Create GraphQL docs (#25)
This creates a GraphQL section for the platform docs.

Part of it is automatized by using the graphql-markdown lib, thus a package.json file was created for the project. This lib handles the GraphQL References section, updated with the results from the Introspection query. Unfortunately, it doesn't create an example query, so the source/graphql.md file was made hybrid by allowing the Reference section to be auto-updated with a single command, while the rest of the content should be manually updated on demand.